requirements.txt
pillow
numpy
PySide6
scikit-learn
imagehash
urllib3
requests
pandas
tflite-runtime; sys_platform == "linux"
